# public/home/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lichi API",
        "version": "0.5"
    },
    "paths": {
        "/api/auth/registration": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация пользователя",
                "operationId": "App\\Http\\Controllers\\AuthController::__construct",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Имя пользоваьтеля",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string32"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Почта пользователя",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string32"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Пароль пользователя",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Авторизует пользователя"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Авторизация пользователя",
                "operationId": "App\\Http\\Controllers\\AuthController::__construct",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Почта пользователя",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string32"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Пароль пользователя",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Авторизует пользователя"
                    }
                }
            }
        },
        "/api/auth/change_password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Изменение пароля пользователя",
                "operationId": "App\\Http\\Controllers\\AuthController::__construct",
                "parameters": [
                    {
                        "name": "new_password",
                        "in": "query",
                        "description": "Новый пароль пользователя",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Изменяет пароль пользователя"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Получение информации о пользователе",
                "operationId": "App\\Http\\Controllers\\AuthController::__construct",
                "responses": {
                    "200": {
                        "description": "Выводит информацию о пользователе"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Получение долгоживущего токена пользователя",
                "operationId": "App\\Http\\Controllers\\AuthController::__construct",
                "responses": {
                    "200": {
                        "description": "Обновляет токен пользователя"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/methods/bots": {
            "get": {
                "tags": [
                    "Bots"
                ],
                "summary": "Вывести всех ботов пользователя",
                "operationId": "App\\Http\\Controllers\\BotController::__construct",
                "responses": {
                    "200": {
                        "description": "Выводит ботов у пользователя"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/methods/bot/{id}": {
            "get": {
                "tags": [
                    "Bots"
                ],
                "summary": "Вывести информацию о боте пользователя",
                "operationId": "App\\Http\\Controllers\\BotController::__construct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id бота",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Выводит информацию о боте пользователя"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bots"
                ],
                "summary": "Изменить бота пользователя",
                "operationId": "App\\Http\\Controllers\\BotController::__construct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id бота",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "storage_id",
                        "in": "query",
                        "description": "id хранилища",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Имя бота",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string32"
                        }
                    },
                    {
                        "name": "vk_status",
                        "in": "query",
                        "description": "Включение (1) или выключение (0) бота Вконтакте",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int1"
                        }
                    },
                    {
                        "name": "tg_status",
                        "in": "query",
                        "description": "Включение (1) или выключение (0) бота Телеграмм",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int1"
                        }
                    },
                    {
                        "name": "vb_status",
                        "in": "query",
                        "description": "Включение (1) или выключение (0) бота Viber",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int1"
                        }
                    },
                    {
                        "name": "vk_token_group",
                        "in": "query",
                        "description": "Токен группы Вконтакте",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string32"
                        }
                    },
                    {
                        "name": "tg_token",
                        "in": "query",
                        "description": "Токен Телеграмм",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string32"
                        }
                    },
                    {
                        "name": "tg_proxy",
                        "in": "query",
                        "description": "Телеграмм-прокси",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string32"
                        }
                    },
                    {
                        "name": "vb_token",
                        "in": "query",
                        "description": "Токен Viber",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновляет информацию о боте пользователя"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bots"
                ],
                "summary": "Удалить бота пользователя",
                "operationId": "App\\Http\\Controllers\\BotController::__construct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id бота",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Удаляет бота пользователя"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/methods/bot": {
            "post": {
                "tags": [
                    "Bots"
                ],
                "summary": "Создать бота пользователя",
                "operationId": "App\\Http\\Controllers\\BotController::__construct",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Имя бота",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "name": "storage_id",
                        "in": "query",
                        "description": "id хранилища пользователя",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Создает бота пользователя"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/methods/bot/{id}/handler": {
            "get": {
                "tags": [
                    "Bots"
                ],
                "summary": "Получает обработчик в виде JSON для бота пользователя",
                "operationId": "App\\Http\\Controllers\\BotController::__construct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id бота",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bots"
                ],
                "summary": "Добавляет или изменяет обработчик бота пользователя",
                "operationId": "App\\Http\\Controllers\\BotController::__construct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id бота",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "handler",
                        "in": "query",
                        "description": "обработчик бота",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string",
                            "items": {
                                "type": "json"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/methods/mailing": {
            "get": {
                "tags": [
                    "Mailing"
                ],
                "summary": "Вывести все рассылки пользователя",
                "operationId": "App\\Http\\Controllers\\MailingController::__construct",
                "responses": {
                    "200": {
                        "description": "Выводит рассылки пользователя"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Mailing"
                ],
                "summary": "Создать рассылку пользователя",
                "operationId": "App\\Http\\Controllers\\MailingController::__construct",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Имя рассылки",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string32"
                        }
                    },
                    {
                        "name": "bot_id",
                        "in": "query",
                        "description": "id бота",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "platform",
                        "in": "query",
                        "description": "Платформа для рассылки [vk, tg, vb]",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Создает рассылку"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/api/methods/mailing/bot/{id}": {
            "get": {
                "tags": [
                    "Mailing"
                ],
                "summary": "Вывести информацию о рассылках бота",
                "operationId": "App\\Http\\Controllers\\MailingController::__construct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id бота",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Выводит информацию о рассылках пользователя"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/api/methods/storages": {
            "get": {
                "tags": [
                    "Storage"
                ],
                "summary": "Вывести все хранилища пользователя",
                "operationId": "App\\Http\\Controllers\\StorageController::__construct",
                "responses": {
                    "200": {
                        "description": "Выводит хранилища у пользователя"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}